requirements-min.txt
beautifulsoup4
polars
httplib2

google-api-python-client
google-auth
google-auth-oauthlib

google-cloud-secret-manager
google-cloud-firestore
google-cloud-storage